# Updates wazero to 1.2.1